requirements.txt
anthropic==0.52.0
python-dotenv==1.1.0
tiktoken==0.9.0
numpy==2.2.4
faiss-cpu==1.11.0
voyageai==0.3.7
sentence-transformers==4.1.0
